xenomai: tune kernel config for low latency
In the Xenomai docs and also on Intel ECI docs, there are comments on
kernel config that help with low latency, which improves the real-time
reliability and performance.

From https://v4.xenomai.org/dovetail/rulesofthumb/:

* CONFIG_DEBUG_IRQ_PIPELINE and CONFIG_DEBUG_DOVETAIL are helpful to
  keep the development process sane.
* CONFIG_RAW_PRINTK is useful and recommended for debugging Xenomai
  out-of-band issues.

From https://v4.xenomai.org/core/user-api/scheduling/:

* CONFIG_EVL_SCHED_QUOTA and CONFIG_EVL_SCHED_TP enabled to give users
  more scheduler options for the Xenomai out-of-band scheduler.

From
https://eci.intel.com/docs/3.1/appendix.html#eci-kernel-configuration-optimizations:

* CONFIG_NO_HZ and CPU_FREQ_STAT disabled to reduce task scheduling
  clock overhead.
* CONFIG_CPU_FREQ_DEFAULT_GOV_PERFORMANCE enabled as default to prevent
  the CPU governor from throttling the CPU frequency for power-saving.
* CONFIG_SUSPEND and CONFIG_PM disabled to prevent power management
  features to introduce spikes to latency.
* CONFIG_VIRT_CPU_ACCOUNTING and CONFIG_VIRT_CPU_ACCOUNTING_GEN for more
  accurate task and CPU time accounting.
* CONFIG_RCU_NOCB_CPU for CPU temporal isolation.

Each config in this commit was enabled individually and tested for 5
minutes under stress, using the Scary Grinder workload as defined on the
doc https://v4.xenomai.org/core/benchmarks/#stress-workloads

A final stress test was run for 3 hours. This is the same idea as the
"EVL stress" test defined on https://v4.xenomai.org/ports/, except it
was not run for 24h.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>