xenomai: add Xenomai 3 in addition to Xenomai 4
On the kernel, Xenomai 3 and 4 use the same base patch set, named
"Dovetail" since the kernel 5.10 and newer.

The difference is the co-kernel: Xenomai 3 uses Cobalt, and Xenomai 4
uses EVL.

Due to this difference, this commit split the current implementation in
two, so users can choose which version of the framework to use.

The kernel is built as recommended on
https://v3.xenomai.org/installation/ and the installation is tested
on a Seeed Studio Odyssey X86J4125 v2 SBC using the official Xenomai
3 tools packaged in a container. Test instructions are available on
https://github.com/leograba/xenomai-torizon-tests

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>