auto-provisioning.sh: allow device ID and name to be overridden
The auto-provisioning script will now source a user defined overrides
file named "auto-provisioning-overrides.sh" that is expected to be found
in /etc/sota or in /usr/lib/sota. The sourced file can define functions
whose output will determine the device ID (get_device_id) and device
name (get_device_name) to be used when registering the device to the
platform.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>