auto-provisioning.service: make it depend on system time sync
Due to the device booting with an older system date, there was a race condition between
auto-provisioning and the service that syncs the system time. If auto-provision would
run first, the certificates would be invalid, because the current date
is prior to theur validity, and the service script would fail.

Closes #90

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>