systemd: Add script to recover from rescue or emergency mode
Several services can trigger rescue or emergency mode on failure.
If either mode is triggered after an OTA update the device will require
manual intervention. This is not desirable for Torizon OS as updates
should be able to recover on their own during failure.

Add a script that will trigger a reboot that will eventually cause a
rollback if emergency or rescue mode is triggered while in the middle
of an update.

Related-to: TOR-3390

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>